warning for missing relation
When the vrops looses connection to vcenter the inventory still will receive return 200 codes assuming everything is correct. However, the child resources below the vcenter are staying empty causing metrics to be missing. This warning should indicate to look into this direction.
TODO: there needs to be an alert to cover this.