[vrops-exporter] Removed custom http code
The custom http code introduced crashes the dashboards and triggers a warning message as most are configured for codes over 500. The problem to be displayed is not really a problem as the resources may not be there and that is fine. We have set the code to 204, which means it is successful but does not contain any content.