catch sigterm in collect
The collector should finish its collect()-run when receiving a SIGTERM. This is mainly used for a use case in kubernetes, where the given pod can be killed which would interrupt the current collect-run. Should be combined with terminationGracePeriodSeconds to avoid k8s waiting forever. A sane value would be something along 120 which is plenty of time to finish gracefully